% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{run}
\usage{
run(country, country_code, start_year = 2013, start_week = 40,
  end_year, end_week = 40, data_deaths, data_ia,
  data_weather = download_weather(country_code),
  data_population = NULL, IArest = TRUE, IAlags = 2, ETlags = 2)
}
\arguments{
\item{country}{a}

\item{country_code}{a}

\item{start_year}{a}

\item{start_week}{a}

\item{end_year}{a}

\item{end_week}{a}

\item{data_deaths}{a}

\item{data_ia}{a}

\item{data_weather}{a}

\item{data_population}{a}

\item{IArest}{a}

\item{IAlags}{a}

\item{ETlags}{a}
}
\description{
run
}
